:sparkles: Refactored and added join team details dialog
