+++ app/tournament/page.tsx
@@ -1,7 +1,7 @@
-import { Button } from '../../components/ui/button';
-import { Card } from '../../components/ui/card';
-import TournamentDetails from './tournament_details';
-import { Separator } from '../../components/ui/separator';
+import { Button } from '@/components/ui/button';
+import { Card } from '@/components/ui/card';
+import TournamentDetails from '../../components/tournament/tournament_details';
+import { Separator } from '@/components/ui/separator';
 import JoinTournamentForm from '../../components/tournament/join_tournament/form';
 import Link from 'next/link';
