+++ components/tournament/join_tournament/tournament_pin_form.tsx
@@ -9,9 +9,9 @@ import {
   FormField,
   FormItem,
   FormMessage,
-} from '../../components/ui/form';
-import { Input } from '../../components/ui/input';
-import { Button } from '../../components/ui/button';
+} from '../../ui/form';
+import { Input } from '../../ui/input';
+import { Button } from '../../ui/button';
 import { Check as CheckIcon } from 'lucide-react';
 import { useEffect } from 'react';
